requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
faiss-cpu==1.7.4
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2

# YouTube and video processing
youtube-transcript-api==0.6.1
yt-dlp==2023.7.6

# MCP tools
# Install from: https://github.com/microsoft/mcp
mcp==0.5.0

# Utilities
numpy==1.24.3
tqdm==4.66.1 